kvx: Support NO_HZ_FULL
Summary:
By default, IRQ work is run from the tick interrupt (see `irq_work_run()` in
`update_process_times()`). When we're in full NOHZ mode, restarting the tick
requires the use of IRQ work and if the only place we run IRQ work is in the
tick interrupt we have an unbreakable cycle. In order to solve this, the
function `arch_irq_work_raise()` sends an IPIs to itself to break this cycle
and get the tick started again. All this is already supported by kvx, it only
needs the function `arch_irq_work_has_interrupt()` to return `true`, to confirm
its support.

Ref T21720

Test Plan: Test a config with NO_HZ_FULL

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T21720

Differential Revision: https://phab.kalray.eu/D32935