kvx qsfp: wait for eeprom cache init using completion
Summary:
In some cases, the driver needs to read some data from the eeprom cache
before the initialization of the latter. The current solution used is to
check the value of the first byte of the cache, but this is not enough.
It does not protect from reading the cache while the cache is being
initialized. Therefore, this commit implements a completion in order to
make sure the eeprom cache is fully initialized before reading it.
eeprom_cache_readb() is removed in order to promote direct cache access.

Ref T21631

Test Plan: CI

Reviewers: O129 linux drivers

Maniphest Tasks: T21631

Differential Revision: https://phab.kalray.eu/D33945