kvx qsfp: poll int flags while IntL deasserted
Summary:
In the driver, the handling of IntL events is implemented with the assumption
that a new interrupt will always be fired while the underlying fault persists.
However, it is not always the case with all cables, such as MFA1A00-C005.
Indeed, according to SFF-8436: "fault bits that are cleared while the underlying
fault persists MAY be immediately set again by the module. This may or may not
cause the IntL to de-assert and then re-assert quickly. Hosts should be tolerant
of both behaviors".
As said above, the implementation should be tolerant. Thus, in this commit,
interrupt flags are polled at regular interval while IntL remains deasserted in
the eeprom. Once it becomes asserted, the interrupt flags are cleared once to
make sure new IntL interrupts can be fired later.

Depends on D33882 D33945

Ref T21631

Test Plan: CI + MFA1A00-C005 cable

Reviewers: O129 linux drivers

Maniphest Tasks: T21631

Differential Revision: https://phab.kalray.eu/D33972