kvx-net: Add global_pfc_config option in sysfs
Summary:
When using DCB to configure PFC, when the link partner switches from
global pause to PFC, we need to know if we should use global PFC or not.
Add a parameter in the sysfs to allow the user to configure this
behavior.

Test Plan: NA

Reviewers: eguillaume, O129 linux drivers

Differential Revision: https://phab.kalray.eu/D34807

(cherry picked from commit 6dbae806dc6d6c1b8bbb22adf69a123c9448a036)