kvx-eth: Fix FOM read in rx_adapt_broadcast
Summary:
The FOM read from RAWLANEX_RX_ADAPT_FOM is incorrect, as it does not
match the FOM values observed in individual lanes when read from
RAWLANE0-3_RX_ADAPT_FOM or RX_SERDES_STATUS registers. Ensure to read
the FOM for all lanes in broadcast mode and report 'ok' only if the FOM
for all lanes is greater than or equal to the threshold.

Ref T22568

Test Plan: manual and CI

Reviewers: O129 linux drivers

Maniphest Tasks: T22568

Differential Revision: https://phab.kalray.eu/D35246