kvx dma: Skip vchan check if no NOC routes present
Summary:
On CV2, both DMA NoC routes and virtual channels are hard-coded for the
ETH TX DMA, yet the DMA DT parsing code requires the vchan field to be
present in the device tree.

Since the code that actually makes use of the vchan value is the same
that also configures the DMA NoC routes, skip the vchan requirement
when no NoC routes are declared in the device tree.

Ref T19776

Test Plan: CI

Reviewers: O39 ethernet, O129 linux drivers, fmalassagne

Reviewed By: O39 ethernet, fmalassagne

Maniphest Tasks: T19776

Differential Revision: https://phab.kalray.eu/D34604