kvx: Fix dame handling
Summary:
Currently there is no DAME handling in the kernel, so a user space
application can cause the kernel to crash. To fix this, the function
`arch_enter_from_user_mode` is called in all exception handlers very
early in the C code, in each `do_xxx` function (via `irqentry_enter()`
or `syscall_enter_from_user_mode()`). So, as long as none of the kernel
code between the actual exception entry and the call to the C code is
triggering a DAME we are fine. We can be sure that the user space was
triggering the given DAME and can send a SIGBUS to the user space
application.
When leaving the kernel the function `arch_exit_to_user_mode_prepare` is
the last function before the asm context restoration code with
interrupts alreay off, so it's safe to call a `barrier` and trigger a
DAME irq here. Although, if this case happens the kernel anway goes into
`panic()` because we are sure the kernel was triggering the DAME.

Ref T21424

Test Plan: CI

Reviewers: O51 Linux Coolidge, jhascoet, ysionneau

Reviewed By: O51 Linux Coolidge, jhascoet, ysionneau

Subscribers: jborne, vray, jhascoet, ysionneau, #linux_coolidge_cc

Maniphest Tasks: T21424

Differential Revision: https://phab.kalray.eu/D35192