irqchip: Add irq-kvx-apic-mailbox driver
The APIC includes a mailbox controller, containing 128 mailboxes.
Each mailbox is an 8-byte word.
Each mailbox can be independently configured with a trigger condition
and an input function.
After a write to a mailbox if the trigger condition is met, an
interrupt will be generated.
Since this hardware block generates IRQs based on writes at some memory
locations, it is both an interrupt-controller and an MSI controller.

Co-developed-by: Clement Leger <clement@clement-leger.fr>
Signed-off-by: Clement Leger <clement@clement-leger.fr>
Co-developed-by: Jules Maselbas <jmaselbas@zdiv.net>
Signed-off-by: Jules Maselbas <jmaselbas@zdiv.net>
Co-developed-by: Julian Vetter <jvetter@kalrayinc.com>
Signed-off-by: Julian Vetter <jvetter@kalrayinc.com>
Co-developed-by: Luc Michel <luc@lmichel.fr>
Signed-off-by: Luc Michel <luc@lmichel.fr>
Signed-off-by: Yann Sionneau <ysionneau@kalrayinc.com>
---

Notes:
V1 -> V2:
- new patch
- removed print on probe success

V2 -> V3: update compatible